🧹 [code health] extract and centralize database helpers
