+++ apps/api/src/routes/__tests__/papers.test.ts
@@ -161,31 +161,7 @@ describe("papers routes", () => {
         expect(body.paper.doi).toBe("10.1234/example");
         expect(body.paper.showViewCount).toBe(true);
         expect(body.paper.publicViewCount).toBe(4);
-        expect(body.organizations).toEqual([]);
         expect(mockDb.batch).toHaveBeenCalledTimes(1);
-    });
-
-    it("GET /api/papers/:id returns organizations only for org_only papers", async () => {
-        const token = await createTestJWT({ sub: "user-1", githubId: "123", name: "Uploader" });
-        mockDb.select = vi
-            .fn()
-            .mockImplementationOnce(() => makeQuery({ getResult: { id: "paper-1", title: "P1", visibility: "org_only", showViewCount: false } }))
-            .mockImplementationOnce(() => makeQuery({ getResult: { id: "user-1" } }))
-            .mockImplementationOnce(() => makeQuery({ allResult: [{ id: "file-1", filename: "paper.pdf" }] }))
-            .mockImplementationOnce(() => makeQuery({ allResult: [{ userId: "user-1", role: "uploader", name: "Uploader", displayName: null, avatarUrl: null }] }))
-            .mockImplementationOnce(() => makeQuery({ allResult: [{ id: "org-1", name: "Org 1", slug: "org-1" }] }));
-
-        const app = await createTestApp();
-        const env = createTestEnv();
-        const res = await app.request(
-            "http://localhost/api/papers/paper-1",
-            { headers: { Authorization: `Bearer ${token}` } },
-            env as any
-        );
-
-        expect(res.status).toBe(200);
-        const body = (await res.json()) as any;
-        expect(body.organizations).toEqual([{ id: "org-1", name: "Org 1", slug: "org-1" }]);
     });
 
     it("POST /api/papers rejects a non-boolean showViewCount", async () => {
@@ -324,7 +300,6 @@ describe("papers routes", () => {
             .fn()
             .mockImplementationOnce(() => makeQuery({ getResult: { id: "paper-1", title: "P1", visibility: "public", showViewCount: false } }))
             .mockImplementationOnce(() => makeQuery({ allResult: [] }))
-            .mockImplementationOnce(() => makeQuery({ allResult: [] }))
             .mockImplementationOnce(() => makeQuery({ allResult: [] }));
 
         const app = await createTestApp();
@@ -404,7 +379,7 @@ describe("papers routes", () => {
         expect(res.status).toBe(403);
     });
 
-    it("PATCH /api/papers/:id requires orgIds when changing visibility to org_only", async () => {
+    it("PATCH /api/papers/:id rejects changing a non-org_only paper to org_only", async () => {
         const token = await createTestJWT({ sub: "user-1", githubId: "123", name: "Uploader" });
         mockDb.select = vi
             .fn()
@@ -428,102 +403,7 @@ describe("papers routes", () => {
 
         expect(res.status).toBe(400);
         const body = (await res.json()) as any;
-        expect(body.error).toContain("orgIds");
-        expect(mockDb.update).not.toHaveBeenCalled();
-    });
-
-    it("PATCH /api/papers/:id updates paperOrgs when changing visibility to org_only", async () => {
-        const token = await createTestJWT({ sub: "user-1", githubId: "123", name: "Uploader" });
-        const where = vi.fn(async () => undefined);
-        const set = vi.fn(() => ({ where }));
-        const deleteWhere = vi.fn(async () => undefined);
-        const insertValues = vi.fn(async () => undefined);
-
-        mockDb.select = vi
-            .fn()
-            .mockImplementationOnce(() => makeQuery({ getResult: { id: "paper-1", visibility: "private" } }))
-            .mockImplementationOnce(() => makeQuery({ getResult: { paperId: "paper-1", userId: "user-1", role: "uploader" } }))
-            .mockImplementationOnce(() => makeQuery({ allResult: [{ orgId: "org-1" }] }))
-            .mockImplementationOnce(() => makeQuery({ allResult: [] })); // existing paperOrgs (none)
-        mockDb.update = vi.fn(() => ({ set }));
-        mockDb.delete = vi.fn(() => ({ where: deleteWhere }));
-        mockDb.insert = vi.fn(() => ({ values: insertValues }));
-
-        const app = await createTestApp();
-        const env = createTestEnv();
-        const res = await app.request(
-            "http://localhost/api/papers/paper-1",
-            {
-                method: "PATCH",
-                headers: {
-                    Authorization: `Bearer ${token}`,
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ visibility: "org_only", orgIds: ["org-1"] }),
-            },
-            env as any,
-        );
-
-        expect(res.status).toBe(200);
-        expect(set).toHaveBeenCalledWith(
-            expect.objectContaining({ visibility: "org_only" }),
-        );
-        expect(deleteWhere).toHaveBeenCalledTimes(1);
-        expect(insertValues).toHaveBeenCalledWith([{ paperId: "paper-1", orgId: "org-1" }]);
-    });
-
-    it("PATCH /api/papers/:id rejects orgIds for non-org visibility with explicit empty array", async () => {
-        const token = await createTestJWT({ sub: "user-1", githubId: "123", name: "Uploader" });
-        mockDb.select = vi
-            .fn()
-            .mockImplementationOnce(() => makeQuery({ getResult: { id: "paper-1", visibility: "private" } }))
-            .mockImplementationOnce(() => makeQuery({ getResult: { paperId: "paper-1", userId: "user-1", role: "uploader" } }));
-
-        const app = await createTestApp();
-        const env = createTestEnv();
-        const res = await app.request(
-            "http://localhost/api/papers/paper-1",
-            {
-                method: "PATCH",
-                headers: {
-                    Authorization: `Bearer ${token}`,
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ visibility: "private", orgIds: [] }),
-            },
-            env as any
-        );
-
-        expect(res.status).toBe(400);
-        expect(await res.json()).toEqual({
-            error: "orgIds can only be specified when visibility is org_only",
-        });
-    });
-
-    it("PATCH /api/papers/:id rejects orgIds when user is not a member", async () => {
-        const token = await createTestJWT({ sub: "user-1", githubId: "123", name: "Uploader" });
-        mockDb.select = vi
-            .fn()
-            .mockImplementationOnce(() => makeQuery({ getResult: { id: "paper-1", visibility: "private" } }))
-            .mockImplementationOnce(() => makeQuery({ getResult: { paperId: "paper-1", userId: "user-1", role: "uploader" } }))
-            .mockImplementationOnce(() => makeQuery({ allResult: [] }));
-
-        const app = await createTestApp();
-        const env = createTestEnv();
-        const res = await app.request(
-            "http://localhost/api/papers/paper-1",
-            {
-                method: "PATCH",
-                headers: {
-                    Authorization: `Bearer ${token}`,
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ visibility: "org_only", orgIds: ["org-1"] }),
-            },
-            env as any,
-        );
-
-        expect(res.status).toBe(403);
+        expect(body.error).toContain("org_only");
         expect(mockDb.update).not.toHaveBeenCalled();
     });
 
@@ -619,45 +499,6 @@ describe("papers routes", () => {
         expect(where).toHaveBeenCalledTimes(1);
     });
 
-    it("PATCH /api/papers/:id clears paperOrgs when leaving org_only", async () => {
-        const token = await createTestJWT({ sub: "user-1", githubId: "123", name: "Uploader" });
-        const updateWhere = vi.fn();
-        const updateSet = vi.fn(() => ({ where: updateWhere }));
-        const deleteWhere = vi.fn();
-        
-        mockDb.select = vi
-            .fn()
-            .mockImplementationOnce(() => makeQuery({ getResult: { id: "paper-1", visibility: "org_only" } }))
-            .mockImplementationOnce(() => makeQuery({ getResult: { paperId: "paper-1", userId: "user-1", role: "uploader" } }));
-        
-        mockDb.update = vi.fn(() => ({ set: updateSet }));
-        mockDb.delete = vi.fn(() => ({ where: deleteWhere }));
-        mockDb.insert = vi.fn();
-        mockDb.batch = vi.fn(async () => undefined);
-
-        const app = await createTestApp();
-        const env = createTestEnv();
-        const res = await app.request(
-            "http://localhost/api/papers/paper-1",
-            {
-                method: "PATCH",
-                headers: {
-                    Authorization: `Bearer ${token}`,
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ visibility: "private" }),
-            },
-            env as any
-        );
-
-        expect(res.status).toBe(200);
-        expect(mockDb.batch).toHaveBeenCalledTimes(1);
-        expect(mockDb.update).toHaveBeenCalledTimes(1);
-        expect(mockDb.delete).toHaveBeenCalledTimes(1);
-        expect(deleteWhere).toHaveBeenCalledTimes(1);
-        expect(mockDb.insert).not.toHaveBeenCalled();
-    });
-
     it("PATCH /api/papers/:id rejects a primitive JSON body", async () => {
         const token = await createTestJWT({ sub: "user-1", githubId: "123", name: "Uploader" });
 
@@ -680,71 +521,6 @@ describe("papers routes", () => {
         const body = (await res.json()) as any;
         expect(body.error).toContain("Invalid JSON body");
         expect(mockDb.select).not.toHaveBeenCalled();
-    });
-
-    it("PATCH /api/papers/:id rejects overlong orgIds", async () => {
-        const token = await createTestJWT({ sub: "user-1", githubId: "123", name: "Uploader" });
-        mockDb.select = vi
-            .fn()
-            .mockImplementationOnce(() => makeQuery({ getResult: { id: "paper-1", visibility: "org_only" } }))
-            .mockImplementationOnce(() => makeQuery({ getResult: { paperId: "paper-1", userId: "user-1", role: "uploader" } }));
-
-        const app = await createTestApp();
-        const env = createTestEnv();
-        const res = await app.request(
-            "http://localhost/api/papers/paper-1",
-            {
-                method: "PATCH",
-                headers: {
-                    Authorization: `Bearer ${token}`,
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ orgIds: ["o".repeat(65)], visibility: "org_only" }),
-            },
-            env as any,
-        );
-
-        expect(res.status).toBe(400);
-        const body = (await res.json()) as any;
-        expect(body.error).toContain("each orgId");
-    });
-
-    it("PATCH /api/papers/:id batches paper and paperOrgs updates when orgIds change", async () => {
-        const token = await createTestJWT({ sub: "user-1", githubId: "123", name: "Uploader" });
-        const deleteWhere = vi.fn(async () => undefined);
-        const insertValues = vi.fn(async () => undefined);
-
-        mockDb.select = vi
-            .fn()
-            .mockImplementationOnce(() => makeQuery({ getResult: { id: "paper-1", visibility: "org_only" } }))
-            .mockImplementationOnce(() => makeQuery({ getResult: { paperId: "paper-1", userId: "user-1", role: "uploader" } }))
-            .mockImplementationOnce(() => makeQuery({ allResult: [{ orgId: "org-1" }] }))
-            .mockImplementationOnce(() => makeQuery({ allResult: [] })); // existing paperOrgs (none)
-        mockDb.delete = vi.fn(() => ({ where: deleteWhere }));
-        mockDb.insert = vi.fn(() => ({ values: insertValues }));
-
-        const app = await createTestApp();
-        const env = createTestEnv();
-        const res = await app.request(
-            "http://localhost/api/papers/paper-1",
-            {
-                method: "PATCH",
-                headers: {
-                    Authorization: `Bearer ${token}`,
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ visibility: "org_only", orgIds: ["org-1"], title: "Updated title" }),
-            },
-            env as any,
-        );
-
-        expect(res.status).toBe(200);
-        expect(mockDb.batch).toHaveBeenCalledTimes(1);
-        expect(mockDb.update).toHaveBeenCalledTimes(1);
-        expect(deleteWhere).toHaveBeenCalledTimes(1);
-        expect(insertValues).toHaveBeenCalledTimes(1);
-        const batchedOps = mockDb.batch.mock.calls[0][0] as unknown[];
-        expect(batchedOps).toHaveLength(3);
     });
 
     it("PATCH /api/papers/:id rejects requests without valid updatable fields", async () => {
