+++ apps/api/src/routes/__tests__/users.test.ts
@@ -467,5 +467,4 @@ describe("users routes", () => {
         const body = (await res.json()) as any;
         expect(body.organizations).toEqual([]);
     });
-
 });
