+++ apps/api/src/routes/collections.ts
@@ -15,13 +15,8 @@ import {
 } from "../db/schema";
 import { authMiddleware } from "../middleware/auth";
 import type { Env, Variables } from "../types";
-import {
-    getOrgBySlug,
-    getOrgMembership,
-    isOrgAdmin,
-    isOrgMember,
-} from "../utils/db";
 import { validateSlug, validateName, validateDescription } from "../utils/validation";
+import { getOrgBySlug, getOrgMembership, isOrgMember, isOrgAdmin } from "../utils/db";
 
 const collectionsRoute = new Hono<{ Bindings: Env; Variables: Variables }>();
 
@@ -38,8 +33,8 @@ function parseVisibility(value: unknown): Visibility | null {
 }
 
 function isUniqueConstraintError(err: unknown): boolean {
-    const message = (err instanceof Error ? err.message : String(err)).toLowerCase();
-    return message.includes("unique") || message.includes("duplicate key");
+    const message = err instanceof Error ? err.message : String(err);
+    return message.includes("UNIQUE") || message.includes("unique") || message.includes("constraint");
 }
 
 async function getCurrentUser(c: any): Promise<CurrentUser> {
@@ -59,478 +54,5 @@ async function getCurrentUser(c: any): Promise<CurrentUser> {
     }
 }
 
-async function canViewCollection(
-    db: ReturnType<typeof drizzle>,
-    collection: typeof collections.$inferSelect,
-    currentUser: CurrentUser,
-): Promise<boolean> {
-    if (collection.visibility === "public") return true;
-    if (!currentUser) return false;
-
-    if (collection.ownerType === "user") {
-        return collection.ownerId === currentUser.id;
-    }
-
-    if (collection.visibility === "private") {
-        return isOrgAdmin(db, collection.ownerId, currentUser.id);
-    }
-
-    return isOrgMember(db, collection.ownerId, currentUser.id);
-}
-
-async function canManageCollection(
-    db: ReturnType<typeof drizzle>,
-    collection: typeof collections.$inferSelect,
-    userId: string,
-): Promise<boolean> {
-    if (collection.ownerType === "user") {
-        return collection.ownerId === userId;
-    }
-    return isOrgAdmin(db, collection.ownerId, userId);
-}
-
-async function canViewPaper(
-    db: ReturnType<typeof drizzle>,
-    paper: { id: string; visibility: string },
-    userId: string | null,
-): Promise<boolean> {
-    if (paper.visibility === "public") return true;
-    if (!userId) return false;
-
-    const isAuthor = await db
-        .select({ paperId: paperAuthors.paperId })
-        .from(paperAuthors)
-        .where(and(eq(paperAuthors.paperId, paper.id), eq(paperAuthors.userId, userId)))
-        .get();
-    if (isAuthor) return true;
-
-    if (paper.visibility === "private") return false;
-
-    const membership = await db
-        .select({ orgId: orgMembers.orgId })
-        .from(orgMembers)
-        .innerJoin(paperOrgs, eq(orgMembers.orgId, paperOrgs.orgId))
-        .where(and(eq(paperOrgs.paperId, paper.id), eq(orgMembers.userId, userId)))
-        .get();
-    return !!membership;
-}
-
-collectionsRoute.post("/collections", authMiddleware, async (c) => {
-    const db = drizzle(c.env.DB);
-    await enableForeignKeys(db);
-
-    let body: any;
-    try {
-        body = await c.req.json();
-    } catch {
-        return c.json({ error: "Invalid JSON body" }, 400);
-    }
-
-    const ownerType = body?.owner_type;
-    if (ownerType !== "user" && ownerType !== "org") {
-        return c.json({ error: "owner_type must be 'user' or 'org'" }, 400);
-    }
-
-    const nameErr = validateName(body?.name);
-    if (nameErr) return c.json({ error: nameErr }, 400);
-
-    const slugErr = validateSlug(body?.slug);
-    if (slugErr) return c.json({ error: slugErr }, 400);
-
-    const descErr = validateDescription(body?.description);
-    if (descErr) return c.json({ error: descErr }, 400);
-
-    const requesterId = c.get("user").sub;
-    const visibility =
-        body?.visibility === undefined ? "private" : parseVisibility(body?.visibility);
-    if (!visibility) {
-        return c.json({ error: "Invalid visibility" }, 400);
-    }
-
-    let ownerId = requesterId;
-    let ownerOrgSlug: string | null = null;
-    if (ownerType === "org") {
-        const inputOrgSlug = typeof body?.org_slug === "string" ? body.org_slug.trim().toLowerCase() : "";
-        const inputOwnerId = typeof body?.owner_id === "string" ? body.owner_id.trim() : "";
-
-        let org = null;
-        if (inputOrgSlug) {
-            org = await getOrgBySlug(db, inputOrgSlug);
-        } else if (inputOwnerId) {
-            org = await db.select().from(orgs).where(eq(orgs.id, inputOwnerId)).get();
-        }
-
-        if (!org) return c.json({ error: "Org not found" }, 404);
-
-        const admin = await isOrgAdmin(db, org.id, requesterId);
-        if (!admin) return c.json({ error: "Forbidden: admin access required" }, 403);
-
-        ownerId = org.id;
-        ownerOrgSlug = org.slug;
-    }
-
-    const id = crypto.randomUUID();
-    const slug = (body.slug as string).trim().toLowerCase();
-
-    try {
-        await db.insert(collections).values({
-            id,
-            ownerType,
-            ownerId,
-            orgSlug: ownerOrgSlug,
-            name: (body.name as string).trim(),
-            slug,
-            description: body?.description ? (body.description as string).trim() : null,
-            visibility,
-        });
-    } catch (err) {
-        if (isUniqueConstraintError(err)) {
-            return c.json({ error: "slug already in use" }, 409);
-        }
-        throw err;
-    }
-
-    const collection = await db.select().from(collections).where(eq(collections.id, id)).get();
-    return c.json({ collection }, 201);
-});
-
-collectionsRoute.get("/collections/:id", async (c) => {
-    const db = drizzle(c.env.DB);
-    const currentUser = await getCurrentUser(c);
-
-    const collection = await db.select().from(collections).where(eq(collections.id, c.req.param("id"))).get();
-    if (!collection) return c.json({ error: "Collection not found" }, 404);
-
-    const visible = await canViewCollection(db, collection, currentUser);
-    if (!visible) return c.json({ error: "Collection not found" }, 404);
-
-    return c.json({ collection });
-});
-
-collectionsRoute.patch("/collections/:id", authMiddleware, async (c) => {
-    const db = drizzle(c.env.DB);
-    await enableForeignKeys(db);
-
-    const id = c.req.param("id");
-    const collection = await db.select().from(collections).where(eq(collections.id, id)).get();
-    if (!collection) return c.json({ error: "Collection not found" }, 404);
-
-    const requesterId = c.get("user").sub;
-    if (!(await canManageCollection(db, collection, requesterId))) {
-        return c.json({ error: "Forbidden" }, 403);
-    }
-
-    let body: any;
-    try {
-        body = await c.req.json();
-    } catch {
-        return c.json({ error: "Invalid JSON body" }, 400);
-    }
-
-    const updates: Record<string, unknown> = {};
-
-    if (body?.name !== undefined) {
-        const e = validateName(body.name);
-        if (e) return c.json({ error: e }, 400);
-        updates.name = (body.name as string).trim();
-    }
-
-    if (body?.slug !== undefined) {
-        const e = validateSlug(body.slug);
-        if (e) return c.json({ error: e }, 400);
-        updates.slug = (body.slug as string).trim().toLowerCase();
-    }
-
-    if (body?.description !== undefined) {
-        const e = validateDescription(body.description);
-        if (e) return c.json({ error: e }, 400);
-        updates.description = body.description ? (body.description as string).trim() : null;
-    }
-
-    if (body?.visibility !== undefined) {
-        if (!VALID_VISIBILITY.includes(body.visibility)) {
-            return c.json({ error: "Invalid visibility" }, 400);
-        }
-        updates.visibility = body.visibility;
-    }
-
-    if (Object.keys(updates).length === 0) {
-        return c.json({ error: "No fields to update" }, 400);
-    }
-
-    try {
-        await db
-            .update(collections)
-            .set({ ...updates, ...touchUpdatedAt() })
-            .where(eq(collections.id, id));
-    } catch (err) {
-        if (isUniqueConstraintError(err)) {
-            return c.json({ error: "slug already in use" }, 409);
-        }
-        throw err;
-    }
-
-    const updated = await db.select().from(collections).where(eq(collections.id, id)).get();
-    return c.json({ collection: updated });
-});
-
-collectionsRoute.delete("/collections/:id", authMiddleware, async (c) => {
-    const db = drizzle(c.env.DB);
-    await enableForeignKeys(db);
-
-    const id = c.req.param("id");
-    const collection = await db.select().from(collections).where(eq(collections.id, id)).get();
-    if (!collection) return c.json({ error: "Collection not found" }, 404);
-
-    const requesterId = c.get("user").sub;
-    if (!(await canManageCollection(db, collection, requesterId))) {
-        return c.json({ error: "Forbidden" }, 403);
-    }
-
-    await db.delete(collections).where(eq(collections.id, id));
-    return c.json({ ok: true });
-});
-
-collectionsRoute.get("/orgs/:slug/collections", async (c) => {
-    const db = drizzle(c.env.DB);
-    const currentUser = await getCurrentUser(c);
-    const org = await getOrgBySlug(db, c.req.param("slug"));
-    if (!org) return c.json({ collections: [] });
-
-    const rows = await db
-        .select()
-        .from(collections)
-        .where(and(eq(collections.ownerType, "org"), eq(collections.ownerId, org.id)))
-        .all();
-
-    let currentUserIsMember = false;
-    let currentUserIsAdmin = false;
-    if (currentUser) {
-        const membership = await getOrgMembership(db, org.id, currentUser.id);
-        currentUserIsMember = !!membership;
-        currentUserIsAdmin = !!membership && (membership.role === "admin" || membership.role === "owner");
-    }
-
-    const filtered = rows.filter((row) => {
-        if (row.visibility === "public") return true;
-        if (!currentUser) return false;
-        if (row.visibility === "private") return currentUserIsAdmin;
-        return currentUserIsMember;
-    });
-
-    return c.json({ collections: filtered });
-});
-
-collectionsRoute.get("/users/:id/collections", async (c) => {
-    const db = drizzle(c.env.DB);
-    const currentUser = await getCurrentUser(c);
-
-    const rows = await db
-        .select()
-        .from(collections)
-        .where(and(eq(collections.ownerType, "user"), eq(collections.ownerId, c.req.param("id"))))
-        .all();
-
-    const filtered = rows.filter((row) => {
-        if (row.visibility === "public") return true;
-        return !!currentUser && currentUser.id === row.ownerId;
-    });
-
-    return c.json({ collections: filtered });
-});
-
-collectionsRoute.post("/collections/:id/papers", authMiddleware, async (c) => {
-    const db = drizzle(c.env.DB);
-    await enableForeignKeys(db);
-
-    const collection = await db.select().from(collections).where(eq(collections.id, c.req.param("id"))).get();
-    if (!collection) return c.json({ error: "Collection not found" }, 404);
-
-    const requesterId = c.get("user").sub;
-    if (!(await canManageCollection(db, collection, requesterId))) {
-        return c.json({ error: "Forbidden" }, 403);
-    }
-
-    let body: any;
-    try {
-        body = await c.req.json();
-    } catch {
-        return c.json({ error: "Invalid JSON body" }, 400);
-    }
-
-    const paperId = typeof body?.paper_id === "string" ? body.paper_id.trim() : "";
-    if (!paperId) return c.json({ error: "paper_id is required" }, 400);
-
-    const paper = await db.select().from(papers).where(eq(papers.id, paperId)).get();
-    if (!paper) return c.json({ error: "Paper not found" }, 404);
-
-    const canView = await canViewPaper(db, paper, requesterId);
-    if (!canView) return c.json({ error: "Paper not found" }, 404);
-
-    const maxOrderRow = await db
-        .select({ maxOrder: sql<number>`COALESCE(MAX(${collectionPapers.sortOrder}), -1)` })
-        .from(collectionPapers)
-        .where(eq(collectionPapers.collectionId, collection.id))
-        .get();
-
-    try {
-        await db.insert(collectionPapers).values({
-            collectionId: collection.id,
-            paperId,
-            sortOrder: (maxOrderRow?.maxOrder ?? -1) + 1,
-        });
-    } catch (err) {
-        if (isUniqueConstraintError(err)) {
-            return c.json({ error: "Paper already added" }, 409);
-        }
-        throw err;
-    }
-    return c.json({ ok: true }, 201);
-});
-
-collectionsRoute.delete("/collections/:id/papers/:paperId", authMiddleware, async (c) => {
-    const db = drizzle(c.env.DB);
-    await enableForeignKeys(db);
-
-    const collection = await db.select().from(collections).where(eq(collections.id, c.req.param("id"))).get();
-    if (!collection) return c.json({ error: "Collection not found" }, 404);
-
-    const requesterId = c.get("user").sub;
-    if (!(await canManageCollection(db, collection, requesterId))) {
-        return c.json({ error: "Forbidden" }, 403);
-    }
-
-    const result = await db
-        .delete(collectionPapers)
-        .where(and(eq(collectionPapers.collectionId, collection.id), eq(collectionPapers.paperId, c.req.param("paperId"))));
-
-    if ((result as any).meta?.changes === 0) {
-        return c.json({ error: "Paper not in collection" }, 404);
-    }
-    return c.json({ ok: true });
-});
-
-collectionsRoute.patch("/collections/:id/papers", authMiddleware, async (c) => {
-    const db = drizzle(c.env.DB);
-    await enableForeignKeys(db);
-
-    const collection = await db.select().from(collections).where(eq(collections.id, c.req.param("id"))).get();
-    if (!collection) return c.json({ error: "Collection not found" }, 404);
-
-    const requesterId = c.get("user").sub;
-    if (!(await canManageCollection(db, collection, requesterId))) {
-        return c.json({ error: "Forbidden" }, 403);
-    }
-
-    let body: any;
-    try {
-        body = await c.req.json();
-    } catch {
-        return c.json({ error: "Invalid JSON body" }, 400);
-    }
-
-    const paperIds = Array.isArray(body?.paper_ids) ? body.paper_ids : [];
-    if (paperIds.length === 0) return c.json({ error: "paper_ids is required" }, 400);
-    if (paperIds.some((v: unknown) => typeof v !== "string" || (v as string).trim() === "")) {
-        return c.json({ error: "paper_ids must be an array of non-empty strings" }, 400);
-    }
-
-    const existingRows = await db
-        .select({ paperId: collectionPapers.paperId })
-        .from(collectionPapers)
-        .where(eq(collectionPapers.collectionId, collection.id))
-        .all();
-
-    const existingSet = new Set(existingRows.map((r) => r.paperId));
-    const uniqueInput = new Set(paperIds);
-    if (uniqueInput.size !== paperIds.length) {
-        return c.json({ error: "paper_ids must not contain duplicates" }, 400);
-    }
-    if (paperIds.length !== existingRows.length) {
-        return c.json({ error: "paper_ids must include all papers in collection" }, 400);
-    }
-    if (paperIds.some((id: string) => !existingSet.has(id))) {
-        return c.json({ error: "paper_ids contains paper not in collection" }, 400);
-    }
-
-    const updateStatements = paperIds.map((pid: string, i: number) =>
-        db
-            .update(collectionPapers)
-            .set({ sortOrder: i })
-            .where(and(eq(collectionPapers.collectionId, collection.id), eq(collectionPapers.paperId, pid))),
-    );
-    await db.batch(updateStatements as [typeof updateStatements[0], ...typeof updateStatements]);
-    return c.json({ ok: true });
-});
-
-collectionsRoute.get("/collections/:id/papers", async (c) => {
-    const db = drizzle(c.env.DB);
-    const currentUser = await getCurrentUser(c);
-
-    const collection = await db.select().from(collections).where(eq(collections.id, c.req.param("id"))).get();
-    if (!collection) return c.json({ error: "Collection not found" }, 404);
-
-    const visible = await canViewCollection(db, collection, currentUser);
-    if (!visible) return c.json({ error: "Collection not found" }, 404);
-
-    const rows = await db
-        .select({
-            id: papers.id,
-            title: papers.title,
-            abstract: papers.abstract,
-            visibility: papers.visibility,
-            year: papers.year,
-            venue: papers.venue,
-            category: papers.category,
-            sortOrder: collectionPapers.sortOrder,
-        })
-        .from(collectionPapers)
-        .innerJoin(papers, eq(collectionPapers.paperId, papers.id))
-        .where(eq(collectionPapers.collectionId, collection.id))
-        .orderBy(asc(collectionPapers.sortOrder))
-        .all();
-
-    const currentUserId = currentUser?.id ?? null;
-    let visiblePapers: typeof rows;
-
-    const restrictedRows = rows.filter((r) => r.visibility !== "public");
-    if (restrictedRows.length === 0) {
-        visiblePapers = rows;
-    } else if (!currentUserId) {
-        visiblePapers = rows.filter((r) => r.visibility === "public");
-    } else {
-        const restrictedIds = restrictedRows.map((r) => r.id);
-
-        const authoredRows = await db
-            .select({ paperId: paperAuthors.paperId })
-            .from(paperAuthors)
-            .where(and(inArray(paperAuthors.paperId, restrictedIds), eq(paperAuthors.userId, currentUserId)))
-            .all();
-        const authoredSet = new Set(authoredRows.map((r) => r.paperId));
-
-        const orgOnlyIds = restrictedRows
-            .filter((r) => r.visibility === "org_only" && !authoredSet.has(r.id))
-            .map((r) => r.id);
-        const orgAccessSet = new Set<string>();
-        if (orgOnlyIds.length > 0) {
-            const orgAccessRows = await db
-                .select({ paperId: paperOrgs.paperId })
-                .from(orgMembers)
-                .innerJoin(paperOrgs, eq(orgMembers.orgId, paperOrgs.orgId))
-                .where(and(inArray(paperOrgs.paperId, orgOnlyIds), eq(orgMembers.userId, currentUserId)))
-                .all();
-            for (const r of orgAccessRows) orgAccessSet.add(r.paperId);
-        }
-
-        visiblePapers = rows.filter((r) => {
-            if (r.visibility === "public") return true;
-            if (authoredSet.has(r.id)) return true;
-            if (r.visibility === "org_only" && orgAccessSet.has(r.id)) return true;
-            return false;
-        });
-    }
-
-    return c.json({ papers: visiblePapers });
-});
 
 export default collectionsRoute;
