+++ apps/api/src/routes/papers.ts
@@ -7,7 +7,6 @@ import {
     paperAuthors,
     paperViews,
     users,
-    orgs,
     coauthorInvites,
     orgMembers,
     paperOrgs,
@@ -35,7 +34,7 @@ const ALLOWED_MIME_TYPES = [
     "image/jpeg",
 ];
 const VALID_FILE_TYPES = ["paper", "slides", "poster", "supplementary"];
-const VALID_VISIBILITY = ["public", "org_only", "private"] as const;
+const VALID_VISIBILITY = ["public", "org_only", "private"];
 const MAX_TITLE_LENGTH = 300;
 const MAX_ABSTRACT_LENGTH = 5000;
 const MAX_LANGUAGE_LENGTH = 32;
@@ -43,14 +42,8 @@ const MAX_EXTERNAL_URL_LENGTH = 2048;
 const MAX_DOI_LENGTH = 255;
 const MAX_VENUE_LENGTH = 255;
 const MAX_TAG_LENGTH = 64;
-const MAX_ORG_ID_LENGTH = 64;
 const VIEW_DEDUPE_WINDOW_MS = 24 * 60 * 60 * 1000;
 const VIEW_STATS_RANGE_DAYS = 30;
-type Visibility = (typeof VALID_VISIBILITY)[number];
-
-function isVisibility(value: string): value is Visibility {
-    return (VALID_VISIBILITY as readonly string[]).includes(value);
-}
 
 function formatDateKey(date: Date): string {
     return date.toISOString().slice(0, 10);
@@ -278,10 +271,9 @@ papersRoute.post("/", authMiddleware, async (c) => {
     )
         return c.json({ error: `title is required (1-${MAX_TITLE_LENGTH} chars)` }, 400);
 
-    const visibilityInput = typeof meta.visibility === "string" ? meta.visibility : "private";
-    if (!isVisibility(visibilityInput))
+    const vis = (meta.visibility as string) || "private";
+    if (!VALID_VISIBILITY.includes(vis))
         return c.json({ error: "Invalid visibility" }, 400);
-    const vis: Visibility = visibilityInput;
 
     const venueType = (meta.venueType as string | null | undefined) ?? null;
     if (venueType !== null && !(VALID_VENUE_TYPES as readonly string[]).includes(venueType))
@@ -303,13 +295,9 @@ papersRoute.post("/", authMiddleware, async (c) => {
         return c.json({ error: "showViewCount must be a boolean" }, 400);
     }
 
-    const orgIdRaw = typeof meta.orgId === "string" ? meta.orgId.trim() : undefined;
-    const orgId = orgIdRaw || undefined;
+    const orgId = meta.orgId as string | undefined;
     if (vis === "org_only" && !orgId) {
         return c.json({ error: "orgId is required for org_only visibility" }, 400);
-    }
-    if (orgId && orgId.length > MAX_ORG_ID_LENGTH) {
-        return c.json({ error: `orgId must be ${MAX_ORG_ID_LENGTH} chars or less` }, 400);
     }
 
     const paperId = crypto.randomUUID();
@@ -563,27 +551,8 @@ papersRoute.get("/:id", async (c) => {
             .where(eq(paperAuthors.paperId, paperId)),
     ])) as [
             (typeof paperFiles.$inferSelect)[],
-            {
-                userId: string;
-                role: string;
-                name: string | null;
-                displayName: string | null;
-                avatarUrl: string | null;
-            }[]
+            { userId: string; role: string; name: string | null; displayName: string | null; avatarUrl: string | null }[]
         ];
-
-    const organizations = paper.visibility === "org_only"
-        ? await db
-            .select({
-                id: orgs.id,
-                name: orgs.name,
-                slug: orgs.slug,
-            })
-            .from(paperOrgs)
-            .innerJoin(orgs, eq(paperOrgs.orgId, orgs.id))
-            .where(eq(paperOrgs.paperId, paperId))
-            .all()
-        : [];
 
     const files = rawFiles.map((file) => ({
         ...file,
@@ -614,7 +583,6 @@ papersRoute.get("/:id", async (c) => {
         },
         files,
         authors,
-        organizations,
     });
 });
 
@@ -1115,12 +1083,6 @@ papersRoute.patch("/:id", authMiddleware, async (c) => {
 
     const updates: Record<string, any> = { ...touchUpdatedAt() };
     let hasRealUpdates = false;
-    if (!isVisibility(paper.visibility)) {
-        return c.json({ error: "Invalid paper visibility state" }, 500);
-    }
-    let nextVisibility: Visibility = paper.visibility;
-    let hasVisibilityInBody = false;
-    let orgIdsFromBody: string[] | undefined;
 
     if ("title" in body) {
         if (typeof body.title !== "string") {
@@ -1145,38 +1107,16 @@ papersRoute.patch("/:id", authMiddleware, async (c) => {
         hasRealUpdates = true;
     }
     if ("visibility" in body) {
-        if (typeof body.visibility !== "string" || !isVisibility(body.visibility)) {
+        if (typeof body.visibility !== "string" || !VALID_VISIBILITY.includes(body.visibility)) {
             return c.json({ error: "Invalid visibility" }, 400);
         }
+        if (body.visibility === "org_only" && paper.visibility !== "org_only") {
+            return c.json(
+                { error: "Changing visibility to org_only is not supported in edit flow" },
+                400,
+            );
+        }
         updates.visibility = body.visibility;
-        nextVisibility = body.visibility;
-        hasVisibilityInBody = true;
-        hasRealUpdates = true;
-    }
-    if ("orgIds" in body) {
-        if (!Array.isArray(body.orgIds)) {
-            return c.json({ error: "orgIds must be an array" }, 400);
-        }
-
-        const normalizedOrgIds: string[] = [];
-        const seen = new Set<string>();
-        for (const orgId of body.orgIds) {
-            if (typeof orgId !== "string") {
-                return c.json({ error: "each orgId must be a string" }, 400);
-            }
-            const trimmed = orgId.trim();
-            if (!trimmed) {
-                return c.json({ error: "orgIds cannot contain empty values" }, 400);
-            }
-            if (trimmed.length > MAX_ORG_ID_LENGTH) {
-                return c.json({ error: `each orgId must be ${MAX_ORG_ID_LENGTH} chars or less` }, 400);
-            }
-            if (seen.has(trimmed)) continue;
-            seen.add(trimmed);
-            normalizedOrgIds.push(trimmed);
-        }
-
-        orgIdsFromBody = normalizedOrgIds;
         hasRealUpdates = true;
     }
     if ("showViewCount" in body) {
@@ -1278,82 +1218,11 @@ papersRoute.patch("/:id", authMiddleware, async (c) => {
         }
         hasRealUpdates = true;
     }
-
-    let shouldReplacePaperOrgs = false;
-    let finalOrgIds: string[] = [];
-
-    if (orgIdsFromBody !== undefined && nextVisibility !== "org_only") {
-        return c.json({ error: "orgIds can only be specified when visibility is org_only" }, 400);
-    }
-
-    if (nextVisibility === "org_only") {
-        // Require explicit organization selection only when switching to org_only
-        // or when orgIds are explicitly edited.
-        if ((hasVisibilityInBody && paper.visibility !== "org_only") || orgIdsFromBody !== undefined) {
-            finalOrgIds = orgIdsFromBody ?? [];
-            if (finalOrgIds.length === 0) {
-                return c.json({ error: "orgIds is required when visibility is org_only" }, 400);
-            }
-
-            const memberships = await db
-                .select({ orgId: orgMembers.orgId })
-                .from(orgMembers)
-                .where(
-                    and(
-                        eq(orgMembers.userId, userId),
-                        inArray(orgMembers.orgId, finalOrgIds),
-                    ),
-                )
-                .all();
-
-            if (memberships.length !== finalOrgIds.length) {
-                return c.json({ error: "Invalid orgIds or not a member" }, 403);
-            }
-
-            const existingOrgs = await db
-                .select({ orgId: paperOrgs.orgId })
-                .from(paperOrgs)
-                .where(eq(paperOrgs.paperId, paperId))
-                .all();
-            
-            const existingOrgIds = existingOrgs.map(o => o.orgId);
-            const finalOrgIdsSet = new Set(finalOrgIds);
-            const existingOrgIdsSet = new Set(existingOrgIds);
-            
-            const areSetsEqual = 
-                finalOrgIdsSet.size === existingOrgIdsSet.size && 
-                [...finalOrgIdsSet].every(id => existingOrgIdsSet.has(id));
-
-            if (!areSetsEqual) {
-                shouldReplacePaperOrgs = true;
-            }
-        }
-    }
-
-    if (!hasRealUpdates && !shouldReplacePaperOrgs) {
+    if (!hasRealUpdates) {
         return c.json({ error: "No valid fields to update" }, 400);
     }
 
-    type BatchItem = Parameters<typeof db.batch>[0][number];
-    const operations: BatchItem[] = [
-        db.update(papers).set(updates).where(eq(papers.id, paperId)),
-    ];
-
-    if (nextVisibility !== "org_only" && hasVisibilityInBody) {
-        operations.push(db.delete(paperOrgs).where(eq(paperOrgs.paperId, paperId)));
-    } else if (shouldReplacePaperOrgs) {
-        operations.push(db.delete(paperOrgs).where(eq(paperOrgs.paperId, paperId)));
-        operations.push(db.insert(paperOrgs).values(
-            finalOrgIds.map((orgId) => ({ paperId, orgId })),
-        ));
-    }
-
-    if (operations.length === 1) {
-        await operations[0];
-    } else {
-        await db.batch(operations as [BatchItem, ...BatchItem[]]);
-    }
-
+    await db.update(papers).set(updates).where(eq(papers.id, paperId));
     return c.json({ ok: true });
 });
