+++ apps/web/src/app/papers/[id]/__tests__/paper-detail-client.test.tsx
@@ -394,82 +394,4 @@ describe("PaperDetailClient", () => {
       await screen.findByText("この論文を閲覧する権限がありません"),
     ).toBeInTheDocument();
   });
-  it("handles image loading failure gracefully", async () => {
-    vi.mocked(apiFetch).mockImplementation(async (input, init) => {
-      const url = String(input);
-      const method = init?.method ?? "GET";
-
-      if (url === "/api/papers/paper-1" && method === "GET") {
-        return jsonResponse({
-          paper: {
-            id: "paper-1",
-            title: "Image Failure Test",
-            abstract: null,
-            visibility: "public",
-            showViewCount: false,
-            publicViewCount: null,
-            externalUrl: null,
-            venue: null,
-            venueType: null,
-            year: null,
-            category: null,
-            tags: null,
-            createdAt: "2026-01-01T00:00:00.000Z",
-            updatedAt: "2026-01-02T00:00:00.000Z",
-          },
-          files: [
-            {
-              id: "file-image",
-              filename: "poster.png",
-              fileType: "poster",
-              sizeBytes: 2048,
-              mimeType: "image/png",
-              downloadUrl: "/api/downloads/poster.png",
-            },
-          ],
-          authors: [
-            {
-              userId: "author-1",
-              role: "author",
-              name: "alice",
-              displayName: "Alice",
-              avatarUrl: null,
-            },
-          ],
-        });
-      }
-
-      if (url === "/api/papers/paper-1/view" && method === "POST") {
-        return jsonResponse({ counted: true });
-      }
-
-      if (url === "/api/papers/paper-1/stats" && method === "GET") {
-        return jsonResponse({
-          totalViews: 0,
-          last7DaysViews: 0,
-          last30DaysViews: 0,
-          dailyViews: [],
-        });
-      }
-
-      if (url.includes("/files/file-image/stream") && method === "GET") {
-        throw new Error("Simulated network failure");
-      }
-
-      throw new Error(`Unexpected request: ${method} ${url}`);
-    });
-
-    render(<PaperDetailClient paperId="paper-1" />);
-
-    await screen.findByRole("heading", { name: "Image Failure Test" });
-
-    await waitFor(() => {
-      expect(console.error).toHaveBeenCalledWith(
-        "Error loading image file-image:",
-        expect.any(Error)
-      );
-    });
-
-    expect(screen.getByText("画像の読み込みに失敗しました")).toBeInTheDocument();
-  });
 });
